completed new desktop design
