--- realtime-chat-v2/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "realtime-chat-v2",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
-    "@vitejs/plugin-react": "^3.1.0",
-    "autoprefixer": "^10.4.13",
-    "postcss": "^8.4.21",
-    "postcss-load-config": "^4.0.1",
-    "postcss-nesting": "^11.2.1",
-    "vite": "^4.1.0"
-  },
-  "postcss": {
-    "plugins": {
-      "postcss-nesting": {},
-      "autoprefixer": {}
-    }
-  }
-}
