--- realtime-chat-v2/src/App.css
@@ -1,18 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    text-decoration: none;
-}
-
-html {
-    height: -webkit-fill-available;
-}
-
-a{
-    color: black;
-}
-
-button{
-    background: none;
-}
