--- realtime-chat-v2/src/App.jsx
@@ -1,16 +1,0 @@
-import './App.css';
-import Modal from './Components/Modal/Modal';
-import Chat from "./Components/Chat/Chat";
-
-function App() {
-
-  return (
-    <>
-      <Modal/>
-      <Chat/>
-    </>
-
-  )
-}
-
-export default App
