--- realtime-chat-v2/src/Components/Chat/Chat.jsx
@@ -1,11 +1,0 @@
-import chatStyles from "./chat.module.css";
-
-const Chat = () =>{
-    return(
-        <div className={chatStyles.chatWrapper}>
-            Lorem ipsum dolor sit amet consectetur adipisicing elit. Atque impedit dolor ipsum asperiores aspernatur veritatis, quia odio hic pariatur eius. Laborum molestiae ducimus quis impedit praesentium obcaecati nesciunt quasi saepe.
-        </div>
-    )
-}
-
-export default Chat;
