--- realtime-chat-v2/src/Components/Chat/chat.module.css
@@ -1,3 +1,0 @@
-.chatWrapper{
-    border: 1px solid black;
-}
