+++ src/Components/Chat/chat.module.css
@@ -1,0 +1,37 @@
+.chatWrapper{
+    display: flex;
+    flex-direction: column;
+    border: 1px solid black;
+    height: 100vh;
+    align-items: center;
+    justify-content: center;
+    background-color: #e66e27;
+    .signOutButtonContainer{
+        display: flex;
+        justify-content: flex-end;
+    }
+    .messageContainer{
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-end;
+        padding: 8px;
+        height: 50vh;
+        border-radius: 5px;
+        background-color: #e2dccc8e;
+        overflow-y: scroll;
+        & input{
+            padding: 8px;
+            border: none;
+            border-radius: 15px;
+        }
+        .myMsg{
+            padding: 8px;
+            margin: 4px;
+            width: fit-content;
+            background-color: whitesmoke;
+            border-radius: 15px;
+        }
+
+    }   
+}
+
