--- src/Components/Chatroom/chatroomStyles.css
@@ -1,83 +1,0 @@
-
-.chatroom-container{
-    display: flex;
-    flex-direction: column;
-    width: fit-content;
-    margin: auto;
-}
-
-.message-list{
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    overflow-y: scroll;
-}
-
-.message-list::-webkit-scrollbar{
-  display: none;
-}
-
-.chat-container{
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  border-radius: 10px;
-  margin: 5px;
-  width: 30vw;
-  height: 70vh;
-  background: rgba(0, 0, 0, 0.7);
-  box-shadow: -10px -9px #e1341e;
-}
-
-.input-button-container{
-  display: flex;
-}
-
-.message-input{
-  padding:8px;
-  border: none;
-}
-
-.send-btn{
-  border: none;
-  background: none;
-  color: #80cee1;
-  position: relative;
-  top: 8px;
-  right: 40px;
-}
-
-.msg{
-  word-wrap: break-word;
-  border: 1px solid black;
-  border-radius: 10px;
-  padding: 8px;
-}
-
-.my-msg{
- align-self: flex-start;
- background-color: #ffffff;
-
-}
-
-.their-msg{
-  align-self: flex-end;
-  word-wrap: break-word;
-  border: 1px solid black;
-  border-radius: 10px;
-  background-color: white;
-
-}
-
-.sign-out-and-options-btn-container{
-  display: flex;
-  justify-content: flex-end;
-  margin: 0 15px 5px 0;
-}
-
-.sign-out-btn{
-  background-color: #80cee1;
-  padding: 5px;
-  border: none;
-}
-
