--- src/Components/Modal/modalStyles.module.css
@@ -1,72 +1,0 @@
-.modalWrapper{
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100vw;
-    height: 100vh;
-    margin: auto;
-    border: 2px solid black;
-    background: rgba(240, 240, 252, 0.582);
-    z-index: 999;
-}
-
-.modal{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    border-radius: 8px;
-    gap: 16px;
-    padding: 8px;
-    background-color: #80cee1;
-}
-
-.loginRegisterContainer{
-  display: flex;
-  justify-content: center;
-  gap: 16px;
-}
-
-.loginContainer{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 8px;
-}
-.registerContainer{
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 8px;
-}
-.modalBtn{
-  position: relative;
-  top: 0px;
-  padding: 15px 25px;
-  border: solid #1a1a1a 3px;
-  border-radius: 20px;
-  margin: 10px;
-  box-shadow: 0px 10px 0px #1a1a1a;
-  background-color: #1ecae1;
-}
-.modalBtn:hover{
-  top: 2px;
-  box-shadow: 0 8px 0 #1a1a1a;
-}
-.modalBtn:active{
-  transition: 200ms ease;
-  top: 8px;
-  box-shadow: 0 2px 0 #1a1a1a;
-  background-color: #80cee1; 
-}
-
-.authContainer{
-  display: flex;
-  align-items: center;
-}
-
-.auth-btn{
-  margin-left: 5px;
-}
